merged main to linh

+++ BE/src/main/java/com/example/demo/controller/ProductDetailController.java
@@ -20,6 +20,6 @@ public class ProductDetailController {
 
     @GetMapping("/product-detail")
     public List<ProductDetail> getProductsByType(@RequestParam int productId) {
-        return productDetailRepository.findByProductId(productId);
+        return productDetailRepository.findByProductId((long) productId);
     }
 }
